requirements.txt
Appium-Python-Client
browserstack-local
browserstack-sdk
percy-selenium==2.1.1
playwright
pytest
pytest-playwright
pytest-variables
pytest-xdist
pytest-base-url
python-dotenv
selenium